Hosted hotspot connect message update.
Small change to the message shown to a user setting up a hotspot they do not personally own. This flow is currently the only way (outside of Kent's unofficial hotspot utility) hotspot hosts can connect their hotspots to Wi-Fi without owner involvement, and the dialog is particularly disorienting to less technically inclined users. Preferably the dialog here would also be able to be closed not backed out of, as the latter intuitively indicates that the network connection did not succeed.