Change Terminology for Light Hotspots
Mirroring changes made to the StatusWidget in Explorer for Light Hotspots, changing "Online" to "Active" and "Needs Attention" to "Inactive".

Justification for the change from "Needs Attention" is that, at least with the current PoC issues, that has been resulting in unnecessary or counterproductive actions, like perpetually power cycling, switching from Ethernet to Wi-Fi, etc. "Inactive" provides the necessary information without implying that action needs to be taken, and it matches what's shown on Explorer.

Note: If this change is merged, the other language files should probably be similarly changed.